Include Lambda ENI cleanup when retrying security group or subnet delete
We've seen cases where delete fails due to timeout. We assume the cause is
lingering resources in the subnet, which would have to be these ENIs.

It seems possible that a Lambda invocation at the wrong time could result in
an ENI popping up between when we search for them and when we try to delete
the subnet or security group.